Mental_health/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
elevenlabs==0.2.26
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
cors==1.0.1
fastapi-cors==0.0.6
